Added functionality: Now your menu looks like this:
1. Balance
2. Add income
3. Do transfer
4. Close account
5. Log out
0. Exit
If the user asks for Balance,  the balance of the account is retrieved from the database and output it into the console.

Add income item allows us to deposit money to the account.

Do transfer item allows transferring money to another account. Following errors are handled:

If the user tries to transfer more money than he/she has, program outputs: Not enough money!
If the user tries to transfer money to the same account, program outputs the following message: You can't transfer money to the same account!
If the receiver's card number doesn’t pass the Luhn algorithm, program outputs: Probably you made a mistake in the card number. Please try again!
If the receiver's card number doesn’t exist, program outputs: Such a card does not exist.
If there is no error, program asks the user how much money they want to transfer and makes the transaction.
If the user chooses the Close an account item, program deletes that account from the database.